Create footer.php
The code inside the footer.php should be included in all the login pages and in all the chatbox files. So instead of copying the code in all the required files, we just direct the file name here. The php include or require function is very helpful I guess. LOL