Create Chat.php
The code inside the Chat.php should be included in all the login file. So instead of copying the code in all the required files, we just direct the file name here. Only to reduce the file of the login.php file and other files which uses this code inside the Chat.php The php include or require function is very helpful I guess. LOL