Update databaseCreate.sql
Use the databaseCreate.sql file to create the appropriate tables only in case the databaseCreator.php file fails to do the intented work. Doing after the php succeeds may result in duplication of the tables in the database. Can be corrected with phpmyadmin or similar service even if you run this file after the databaseCreator.php